pyproject.toml
[project]
name = "extraction"
version = "0.0.1"
description = "Multi-format document extraction library for EPUB, PDF, HTML, Markdown, and JSON documents"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {email = "hello-world-bfree@pm.me"}
]
keywords = ["extraction", "epub", "pdf", "html", "markdown", "document-processing", "chunking", "rag", "embeddings", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Text Processing :: Markup",
    "Topic :: Text Processing :: Linguistic",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
]

[project.urls]
Homepage = "https://github.com/hello-world-bfree/extraction"
Documentation = "https://hello-world-bfree.github.io/extraction/"
Repository = "https://github.com/hello-world-bfree/extraction"

[project.optional-dependencies]
pdf = [
    "pdfplumber>=0.10.0",
]
vatican = [
    "boto3>=1.28.0",
    "botocore>=1.31.0",
]
scraping = [
    "playwright>=1.42.0",
    "Pillow>=10.0.0",
]
epub-images = [
    "Pillow>=10.0.0",
]
s3 = [
    "boto3>=1.34.0",
]
images = [
    "extraction[scraping,epub-images,s3]",
]
finetuning = [
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "sentencepiece>=0.1.99",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
extract = "extraction.cli.extract:main"
vatican-extract = "extraction.cli.vatican_extract:main"
token-rechunk = "extraction.tools.token_rechunker:main"
extract-images = "extraction.cli.scrape_images:main"

[build-system]
requires = ["uv>=0.5.11,<0.6.0"]
build-backend = "uv"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "regression: marks tests as regression tests for previously fixed bugs",
]
addopts = "-v --strict-markers"

[tool.coverage.run]
source = ["src/extraction"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
